Fixed overflow errors in relations with y^2 (1-x^2)=(x^2+2 y-1)^2

+++ Trunk/Source/Source/DrawThread.h
@@ -61,8 +61,8 @@ private:
   void DrawLabel(TTextLabel &Label);
   void DrawRelation(const TRelation &Relation);
   void EquationLoop(TRelation &Relation, std::vector<TRect> &Points, bool Loop);
-  static bool CheckResult1(const double Result[3]);
-  static bool CheckResult2(const double Result[3]);
+  static bool CheckResult1(const long double Result[3]);
+  static bool CheckResult2(const long double Result[3]);
 
   //Draw elements
   void Visit(TBaseFuncType &Func);
